requirements.txt
django
djangorestframework
markdown
django-filter
django-rest-swagger
djangorestframework-simplejwt
django-cors-headers

djangorestframework-jwt
django-rest-auth 
django-allauth